Refactor the save popup code and test it.
There were four identical implementations of the `popup_save` function
that were all not tested.  This has been refactored into one with a
simple test.
